+++ CHANGELOG.md
@@ -17,6 +17,8 @@
   ([#6866](https://github.com/mitmproxy/mitmproxy/pull/6866), @mhils)
 * Fix slowdown when sending large data over HTTP/2
   ([#6875](https://github.com/mitmproxy/mitmproxy/pull/6875), @aib)
+* Add an option to strip HTTPS records from DNS responses to block encrypted ClientHellos.
+  ([#6876](https://github.com/mitmproxy/mitmproxy/pull/6876), @errorxyz)
 
 
 ## 17 April 2024: mitmproxy 10.3.0
@@ -30,7 +32,7 @@
 * Fix multipart form content view being unusable.
   ([#6653](https://github.com/mitmproxy/mitmproxy/pull/6653), @DaniElectra)
 * Documentation Improvements on CA Certificate Generation
-  ([#5370](https://github.com/mitmproxy/mitmproxy/pull/5370), @zioalex)  
+  ([#5370](https://github.com/mitmproxy/mitmproxy/pull/5370), @zioalex)
 * Make it possible to read flows from stdin with mitmweb.
   ([#6732](https://github.com/mitmproxy/mitmproxy/pull/6732), @jaywor1)
 * Update aioquic dependency to >= 1.0.0, < 2.0.0.